relearn-worker/requirements.txt
boto3
pika
keras
pillow